* kparse.c (strutol): Cast argument to isupper()/tolower() to int
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13368 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb4/realmofhost.c
@@ -104,7 +104,7 @@ krb_realmofhost(host)
 		ret_realm[REALM_SZ] = '\0';
 		/* Upper-case realm */
 		for (cp = ret_realm; *cp; cp++)
-			if (islower(*cp))
+			if (islower((int) (*cp)))
 				*cp = toupper(*cp);
 	} else {
 		krb_get_lrealm(ret_realm, 1);
